refactor(manage): migrate entra-groups utility to TypeScript and update imports
