--- apps/manage/src/util/entra-groups-types.d.ts
@@ -1,6 +1,0 @@
-import type { GroupMember } from '@pins/crowndev-lib/graph/types';
-
-export interface EntraGroupMembers {
-	caseOfficers: GroupMember[];
-	inspectors: GroupMember[];
-}
